efactor(bff): restructure graphql code into domain folders
