+++ packages/bff/src/graphql/nexusTypes.ts
@@ -1,0 +1,23 @@
+export * from './shared/types.ts';
+
+export * from './profile/types.ts';
+export * from './profile/queries.ts';
+export * from './profile/mutations.ts';
+
+export * from './notifications/types.ts';
+export * from './notifications/queries.ts';
+export * from './notifications/mutations.ts';
+
+export * from './savedSearches/types.ts';
+export * from './savedSearches/queries.ts';
+export * from './savedSearches/mutations.ts';
+
+export * from './organizations/types.ts';
+export * from './organizations/queries.ts';
+
+export * from './serviceResources/types.ts';
+export * from './serviceResources/queries.ts';
+
+export * from './username/types.ts';
+export * from './username/queries.ts';
+export * from './username/mutations.ts';
