--- packages/bff/src/graphql/types/index.ts
@@ -1,7 +1,0 @@
-export * from './profile.ts';
-export * from './savedSearches.ts';
-export * from './organization.ts';
-export * from './serviceResources/serviceResources.ts';
-export * from './serviceResources/dpServiceResources.ts';
-export * from './mutation.ts';
-export * from './query.ts';
