--- packages/bff/src/graphql/types/mutation.ts
@@ -1,422 +1,0 @@
-import { logger } from '@altinn/dialogporten-node-logger';
-import { booleanArg, extendType, intArg, nonNull, stringArg } from 'nexus';
-import config from '../../config.js';
-import { decryptPersonUrn } from '../../party/personUrnCipher.ts';
-import {
-  addFavoriteParty,
-  addFavoritePartyToGroup,
-  deleteFavoriteParty,
-  deleteNotificationsSetting,
-  getOrCreateProfile,
-  sendVerificationCode,
-  setPreSelectedParty,
-  updateLanguageInCore,
-  updateNotificationsSetting,
-  updateSIPrivatePhoneNumber,
-  updateShowClientUnits,
-  updateShowDeletedEntities,
-  verifyAddress,
-} from '../functions/profile.ts';
-import { setUsername } from '../functions/register.ts';
-import { createSavedSearch, deleteSavedSearch, updateSavedSearch } from '../functions/savedsearch.ts';
-import { languageCodes, updateAltinnPersistentContextValue } from './cookie.js';
-import {
-  NotificationSettingsInput,
-  Response,
-  SavedSearchInput,
-  SavedSearches,
-  SendVerificationCodeInput,
-  VerifyAddressInput,
-} from './index.ts';
-
-export const Mutation = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('deleteSavedSearch', {
-      type: Response,
-      args: {
-        id: nonNull(intArg()),
-      },
-      resolve: async (_, args) => {
-        const { id } = args;
-        try {
-          const result = await deleteSavedSearch(id);
-          return { success: result?.affected && result?.affected > 0, message: 'Saved search deleted successfully' };
-        } catch (error) {
-          logger.error(error, 'Failed to delete saved search:');
-          return { success: false, message: 'Failed to delete saved search' };
-        }
-      },
-    });
-  },
-});
-
-export const UpdateSavedSearch = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('updateSavedSearch', {
-      type: Response,
-      args: {
-        id: nonNull(intArg()),
-        name: stringArg(),
-      },
-      resolve: async (_, args) => {
-        const { id, name } = args;
-        try {
-          await updateSavedSearch(id, name);
-          return { success: true, message: 'Saved search updated successfully' };
-        } catch (error) {
-          logger.error(error, 'Failed to updated saved search:');
-          return { success: false, message: 'Failed to updated saved search' };
-        }
-      },
-    });
-  },
-});
-
-export const CreateSavedSearch = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('createSavedSearch', {
-      type: SavedSearches,
-      args: {
-        name: stringArg(),
-        data: SavedSearchInput,
-      },
-      resolve: async (_, { name, data }, ctx) => {
-        try {
-          const profile = await getOrCreateProfile(ctx);
-          if (!profile) {
-            throw new Error('Profile not found or could not be created');
-          }
-          if (!data) {
-            throw new Error('Data are required to create a saved search');
-          }
-          const resolvedData = {
-            ...data,
-            urn: Array.isArray(data.urn) ? data.urn.map((u: string) => decryptPersonUrn(u) as string) : data.urn,
-          };
-          return await createSavedSearch({ name, data: resolvedData, profile });
-        } catch (error) {
-          logger.error(error, 'Failed to create saved search:');
-          return error;
-        }
-      },
-    });
-  },
-});
-
-export const AddFavoriteParty = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('addFavoriteParty', {
-      type: Response,
-      args: {
-        partyId: stringArg(),
-      },
-      resolve: async (_, { partyId }, ctx) => {
-        try {
-          await addFavoriteParty(ctx, partyId);
-          return { success: true, message: 'FavoriteParty added successfully' };
-        } catch (error) {
-          logger.error(error, 'Failed to add favorite party:');
-          return error;
-        }
-      },
-    });
-  },
-});
-
-export const AddFavoritePartyToGroup = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('addFavoritePartyToGroup', {
-      type: Response,
-      args: {
-        partyId: stringArg(),
-        groupName: stringArg(),
-      },
-      resolve: async (_, { partyId, groupName }, ctx) => {
-        try {
-          await addFavoritePartyToGroup(ctx.session.get('pid'), partyId, groupName);
-          return { success: true, message: 'FavoriteParty added successfully' };
-        } catch (error) {
-          logger.error(error, 'Failed to add favorite party:');
-          return error;
-        }
-      },
-    });
-  },
-});
-
-export const UpdateNotificationSetting = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('updateNotificationSetting', {
-      type: Response,
-      args: {
-        data: NotificationSettingsInput,
-      },
-      resolve: async (_, { data }, ctx) => {
-        try {
-          const response = await updateNotificationsSetting(data, ctx);
-          if (typeof response !== 'undefined') {
-            return { success: true, message: 'NotificationSetting updated successfully' };
-          }
-          return { success: false, message: 'NotificationSetting updated failed' };
-        } catch (error) {
-          logger.error(error, 'Failed to update NotificationSetting:');
-          return error;
-        }
-      },
-    });
-  },
-});
-
-export const DeleteNotificationSetting = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('deleteNotificationSetting', {
-      type: Response,
-      args: {
-        partyUuid: stringArg(),
-      },
-      resolve: async (_, { partyUuid }, ctx) => {
-        try {
-          await deleteNotificationsSetting(partyUuid, ctx);
-          return { success: true, message: 'NotificationSetting deleted successfully' };
-        } catch (error) {
-          logger.error(error, 'Failed to delete NotificationSetting:');
-          return error;
-        }
-      },
-    });
-  },
-});
-
-export const DeleteFavoriteParty = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('deleteFavoriteParty', {
-      type: Response,
-      args: {
-        partyId: stringArg(),
-      },
-      resolve: async (_, { partyId }, ctx) => {
-        try {
-          await deleteFavoriteParty(ctx, partyId);
-          return { success: true, message: 'Favorite Party deleted successfully' };
-        } catch (error) {
-          logger.error(error, 'Failed to delete favorite party:');
-          return error;
-        }
-      },
-    });
-  },
-});
-
-export const UpdateLanguage = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('updateLanguage', {
-      type: Response,
-      args: {
-        language: stringArg(),
-      },
-      resolve: async (_, { language }, ctx) => {
-        try {
-          await updateLanguageInCore(ctx, language);
-          const ul = languageCodes[language];
-
-          if (ul) {
-            const current = ctx.request.raw.cookies?.altinnPersistentContext;
-            const value = updateAltinnPersistentContextValue(current, ul);
-            ctx.request.context.reply.setCookie('altinnPersistentContext', value, {
-              path: '/',
-              expires: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30),
-              httpOnly: true,
-              secure: true,
-              domain: config.authContextCookieDomain,
-              encode: (v: string) => v,
-            });
-          }
-
-          return { success: true };
-        } catch (error) {
-          logger.error(error, 'Failed to update language:');
-          return error;
-        }
-      },
-    });
-  },
-});
-
-export const setShouldShowSubEntities = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('setShouldShowSubEntities', {
-      type: Response,
-      args: {
-        shouldShowDeletedEntities: booleanArg(),
-      },
-      resolve: async (_, { shouldShowDeletedEntities }, ctx) => {
-        try {
-          await updateShowDeletedEntities(ctx, shouldShowDeletedEntities);
-          return { success: true };
-        } catch (error) {
-          logger.error(error, 'Failed to update profile setting preference:');
-          return { success: false, message: 'Failed to update profile setting preference' };
-        }
-      },
-    });
-  },
-});
-
-export const setShowClientUnits = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('setShowClientUnits', {
-      type: Response,
-      args: {
-        shouldShowClientUnits: booleanArg(),
-      },
-      resolve: async (_, { shouldShowClientUnits }, ctx) => {
-        try {
-          await updateShowClientUnits(ctx, shouldShowClientUnits);
-          return { success: true };
-        } catch (error) {
-          logger.error(error, 'Failed to update profile setting preference:');
-          return { success: false, message: 'Failed to update profile setting preference' };
-        }
-      },
-    });
-  },
-});
-
-/* This endpoint is responsible for both sending and resending confirmation code */
-export const VerifyAddress = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('verifyAddress', {
-      type: Response,
-      args: {
-        data: VerifyAddressInput,
-      },
-      resolve: async (_, { data }, ctx) => {
-        try {
-          return await verifyAddress(data, ctx);
-        } catch (error) {
-          logger.error(error, 'Failed to verify address:');
-          return { success: false, message: 'Failed to verify address' };
-        }
-      },
-    });
-  },
-});
-
-export const SendVerificationCode = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('sendVerificationCode', {
-      type: Response,
-      args: {
-        data: SendVerificationCodeInput,
-      },
-      resolve: async (_, { data }, ctx) => {
-        try {
-          return await sendVerificationCode(data, ctx);
-        } catch (error) {
-          logger.error(error, 'Failed to send verification code:');
-          return { success: false, message: 'Failed to send verification code' };
-        }
-      },
-    });
-  },
-});
-
-export const UpdateSIPrivatePhoneNumber = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('updateSIPrivatePhoneNumber', {
-      type: Response,
-      args: {
-        value: stringArg(),
-      },
-      resolve: async (_, { value }, ctx) => {
-        const result = await updateSIPrivatePhoneNumber(value ?? null, ctx);
-        if (result?.success) {
-          return { success: true, message: 'Phone number updated successfully' };
-        }
-        return { success: false, message: 'Failed to update phone number' };
-      },
-    });
-  },
-});
-
-const USERNAME_PATTERN = /^[a-z][a-z0-9._@-]{5,63}$/i;
-const NATIONAL_IDENTITY_NUMBER_PATTERN = /^\d{11}$/;
-
-export const SetUsername = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('setUsername', {
-      type: Response,
-      // username: null clears the current username
-      args: { username: stringArg() },
-      resolve: async (_, { username }, ctx) => {
-        const pid = ctx.session.get('pid');
-        if (typeof pid !== 'string' || !pid) {
-          return { success: false, message: 'No pid in session' };
-        }
-
-        // This functionality is however not supported in frontend
-        if (!NATIONAL_IDENTITY_NUMBER_PATTERN.test(pid)) {
-          return { success: false, message: 'Setting username is not supported for this user' };
-        }
-
-        if (username !== null && username !== undefined && !USERNAME_PATTERN.test(username)) {
-          return {
-            success: false,
-            message:
-              'Username must be 6-64 characters long, start with a letter, and can only contain letters, digits, dots, underscores, hyphens or @ signs.',
-          };
-        }
-
-        const party = `urn:altinn:person:identifier-no:${pid}`;
-
-        try {
-          return await setUsername(party, username ?? null);
-        } catch (error) {
-          logger.error(error, 'Failed to set username:');
-          return { success: false, message: 'Failed to set username' };
-        }
-      },
-    });
-  },
-});
-
-export type PreselectedPartyOperationType = 'set' | 'unset';
-export const SetPreSelectedParty = extendType({
-  type: 'Mutation',
-  definition(t) {
-    t.field('setPreSelectedParty', {
-      type: Response,
-      args: { partyUuid: stringArg(), operationType: nonNull(stringArg()) },
-      resolve: async (_, { partyUuid, operationType }, ctx) => {
-        if (!partyUuid) {
-          return { success: false, message: 'partyUuid is required' };
-        }
-        if (operationType !== 'set' && operationType !== 'unset') {
-          return { success: false, message: 'operationType must be "set" or "unset"' };
-        }
-        try {
-          await setPreSelectedParty(ctx, partyUuid, operationType as PreselectedPartyOperationType);
-          return { success: true, message: 'PreSelectedParty set successfully' };
-        } catch (error) {
-          logger.error(error, 'Failed to set preselected party:');
-          return { success: false, message: 'Failed to set preselected party' };
-        }
-      },
-    });
-  },
-});
