+++ packages/bff/src/graphql/username/types.ts
@@ -1,0 +1,8 @@
+import { objectType } from 'nexus';
+
+export const UsernameResponse = objectType({
+  name: 'UsernameResponse',
+  definition(t) {
+    t.nullable.string('username');
+  },
+});
